Added resource files for translations Converted component to view component with model
