+++ src/EPR.RegulatorService.Frontend.Web/ViewModels/RegistrationSubmissions/RegistrationSubmissionComment.cs
@@ -1,0 +1,7 @@
+namespace EPR.RegulatorService.Frontend.Web.ViewModels.RegistrationSubmissions;
+
+public class RegistrationSubmissionComment
+{
+    public string Title { get; set; }
+    public string Comment { get; set; }
+}
